pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pragmatic"
version = "0.1.0"
description = "A composable agent framework for Python"
requires-python = ">=3.11"
license = "MIT"
dependencies = ["pyyaml>=6.0", "openrouter>=0.8", "python-dotenv>=1.0"]

[project.scripts]
pragmatic = "pragmatic.__main__:main"

[project.optional-dependencies]
test = ["tox>=4.0", "pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["pragmatic*"]

[project.urls]
Repository = "https://github.com/SquareRoundCurly/pragmatic"
